// src/types/likedPosts.ts
export interface LikedPostItem {
  postId?: number
  postImageUrl: string
  postCategory: string
  postStatus: string
  postTitle: string
  postDescription: string
  requestCreatedAt: string
  profileImageUrl: string
  postMemberNickname: string
  price: number
  priceUnit: string
  likeCount: number
}